# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  SubVolumeGroupInfo:
    properties:
      atime:
        type: string
      bytes_pcent:
        type: string
      bytes_quota:
        type: integer
      bytes_used:
        type: integer
      created_at:
        type: string
      ctime:
        type: string
      data_pool:
        type: string
      gid:
        type: integer
      mode:
        type: integer
      mon_addrs:
        items:
          type: string
        type: array
      mtime:
        type: string
      uid:
        type: integer
    type: object
  SubVolumeGroupList:
    properties:
      info:
        $ref: '#/definitions/SubVolumeGroupInfo'
      name:
        type: string
      path:
        type: string
      snapshot:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    email: support@ablecloud.io
    name: 윤여천
    url: http://www.ablecloud.io
  description: This is a GlueAPI server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Glue-API
  version: v1.0
paths:
...
  /api/v1/gluefs/subvolume/group:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS 볼륨의 그룹을 삭제합니다.
      parameters:
      - description: Glue FS Volume Name
        in: query
        name: vol_name
        required: true
        type: string
      - description: Glue FS Volume Group Name
        in: query
        name: group_name
        required: true
        type: string
      - description: Glue FS Volume Group Path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Delete of Glue FS Volume Group
      tags:
      - GlueFS-SubVolume-Group
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS볼륨의 그룹에 대한 상세 정보 및 리스트를 보여줍니다.
      parameters:
      - description: Glue FS Volume Name
        in: query
        name: vol_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SubVolumeGroupList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Detail Info and List of Glue FS Volume Groups
      tags:
      - GlueFS-SubVolume-Group
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS 볼륨의 그룹을 생성합니다.
      parameters:
      - description: Glue FS Volume Name
        in: formData
        name: vol_name
        required: true
        type: string
      - description: Glue FS Volume Group Name
        in: formData
        name: group_name
        required: true
        type: string
      - description: Glue FS Volume Group Size(default GB)
        in: formData
        name: size
        required: true
        type: integer
      - description: Glue FS Volume Group Data Pool Name
        in: formData
        name: data_pool_name
        required: true
        type: string
      - description: Glue FS Volume Group Permissions
        in: formData
        name: mode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create of Glue FS Volume Group
      tags:
      - GlueFS-SubVolume-Group
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: GlueFS 볼륨의 그룹의 할당된 사이즈를 수정합니다.
      parameters:
      - description: Glue FS Volume Name
        in: formData
        name: vol_name
        required: true
        type: string
      - description: Glue FS Volume Group Name
        in: formData
        name: group_name
        required: true
        type: string
      - description: Glue FS Volume Group New Size(default GB)
        in: formData
        name: new_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Update Size of Glue FS Volume Group
      tags:
      - GlueFS-SubVolume-Group
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"